int32_t -> size_t

+++ source/extensions/transport_sockets/tls/cert_validator/default_validator.cc
@@ -473,7 +473,7 @@ Envoy::Ssl::CertificateDetailsPtr DefaultCertValidator::getCaCertInformation() c
   return Utility::certificateDetails(ca_cert_.get(), getCaFileName(), time_source_);
 }
 
-absl::optional<int32_t> DefaultCertValidator::daysUntilFirstCertExpires() const {
+absl::optional<size_t> DefaultCertValidator::daysUntilFirstCertExpires() const {
   return Utility::getDaysUntilExpiration(ca_cert_.get(), time_source_);
 }
